+++ controller/konnect/ops/ops.go
@@ -65,7 +65,7 @@ func Create[
 	case *konnectv1alpha1.KonnectCloudGatewayNetwork:
 		err = createKonnectNetwork(ctx, sdk.GetCloudGatewaysSDK(), ent)
 	case *konnectv1alpha1.KonnectCloudGatewayDataPlaneGroupConfiguration:
-		err = createKonnectDataPlaneGroupConfiguration(ctx, sdk.GetCloudGatewaysSDK(), ent, sdk.GetServer().Region())
+		err = createKonnectDataPlaneGroupConfiguration(ctx, sdk.GetCloudGatewaysSDK(), cl, ent, sdk.GetServer().Region())
 	case *configurationv1alpha1.KongService:
 		err = createService(ctx, sdk.GetServicesSDK(), ent)
 	case *configurationv1alpha1.KongRoute:
@@ -318,7 +318,7 @@ func Delete[
 
 	// Clear the instance field from the error to avoid requeueing the resource
 	// because of the trace ID in the instance field is different for each request.
-	err = clearInstanceFromError(err)
+	err = ClearInstanceFromError(err)
 
 	return err
 }
@@ -398,7 +398,7 @@ func Update[
 	case *konnectv1alpha1.KonnectCloudGatewayNetwork:
 		err = updateKonnectNetwork(ctx, sdk.GetCloudGatewaysSDK(), ent)
 	case *konnectv1alpha1.KonnectCloudGatewayDataPlaneGroupConfiguration:
-		err = updateKonnectDataPlaneGroupConfiguration(ctx, sdk.GetCloudGatewaysSDK(), ent, sdk.GetServer())
+		err = updateKonnectDataPlaneGroupConfiguration(ctx, sdk.GetCloudGatewaysSDK(), cl, ent, sdk.GetServer())
 	case *configurationv1alpha1.KongService:
 		err = updateService(ctx, sdk.GetServicesSDK(), ent)
 	case *configurationv1alpha1.KongRoute:
@@ -625,11 +625,11 @@ func getMatchingEntryFromListResponseData[
 	return id, nil
 }
 
-// clearInstanceFromError clears the instance field from the error.
+// ClearInstanceFromError clears the instance field from the error.
 // This is needed because the instance field contains the trace ID which changes
 // with each request and makes the reconciliation loop requeue the resource
 // instead of performing the backoff.
-func clearInstanceFromError(err error) error {
+func ClearInstanceFromError(err error) error {
 	var errBadRequest *sdkkonnecterrs.BadRequestError
 	if errors.As(err, &errBadRequest) {
 		errBadRequest.Instance = ""
@@ -640,6 +640,12 @@ func clearInstanceFromError(err error) error {
 	if errors.As(err, &errConflict) {
 		errConflict.Instance = ""
 		return errConflict
+	}
+
+	var errNotFound *sdkkonnecterrs.NotFoundError
+	if errors.As(err, &errNotFound) {
+		errNotFound.Instance = ""
+		return errNotFound
 	}
 
 	return err